fix(hol-guard): close drawer before intercept proof modal
Collapse intercept proof parsing into the shared action-result module and
close the manager drawer when a test opens the proof modal.
